Created Module for colour Add Module Remove body onload Add window.onload add HTML ref to module Add HTML type=module tags
